refactor: making the eula widgets private
