+++ packages/ubuntu_provision/lib/eula.dart
@@ -1,3 +1,2 @@
 export 'src/eula/eula_l10n.dart';
 export 'src/eula/eula_page.dart';
-export 'src/eula/eula_widgets.dart';
